core: targeted operations
Add `-target=resource` flag to core operations, allowing users to
target specific resources in their infrastructure. When `-target` is
used, the operation will only apply to that resource and its
dependencies.

The calculated dependencies are different depending on whether we're
running a normal operation or a `terraform destroy`.

Generally, "dependencies" refers to resources falling _before_ the
target in the graph, because their changes are required to accurately
act on the target.

For destroys, "dependencies" are those resources which fall _after_ the
target. These resources depend on our target, which is going to be
destroyed, so they should also be destroyed.